Lots of changes, additions to client.
Working on adding models to client.  Event system to come shortly.

@TODO: Client message config needs to be taken out of client side.  Only need a more generalized settings module for
UI options and other random junk.  Only server really 'needs' validation of messages. But the idea being that there is some
security benefit.  So if there is none to be gained, then it can just be scrapped and command modules loaded in a simple manner.